+++ gfx/font_d3d_driver.h
@@ -44,7 +44,7 @@ extern d3d_font_renderer_t d3d_win32_font;
 
 bool d3d_font_init_first(
       const void **font_driver, void **font_handle,
-      void *video_data, const char *font_path, unsigned font_size);
+      void *video_data, const char *font_path, float font_size);
 
 #ifdef __cplusplus
 }
